Adding sound to click

+++ app/src/main/java/com/ocr/john/omood/model/MyViewHolder.java
@@ -52,7 +52,7 @@ public  class MyViewHolder extends RecyclerView.ViewHolder implements View.OnCli
 
         Log.i(BUNDLE_VH_INFO,"Apply click on image with tag and current position : " + getAdapterPosition() );
 
-        vhListener.onViewHolderClick(view,getAdapterPosition());
+        vhListener.onViewHolderClick(view,(short)getAdapterPosition());
     }
 
 }
